included backend for upload,download and created redirection for thankyou page to download the content using link

+++ func.php
@@ -3,21 +3,69 @@
   require_once('connection.php');
 
 
+
+function getFilename($broadcastchannel ,$roomnumber)
+{
+  $conn = mysqli_connect('localhost', 'root', '','instantsharing');
+ $sql = "SELECT filename FROM storage WHERE broadcastchannel= '$broadcastchannel' and roomnumber = '$roomnumber'";
+ 
+ $filename =  null;
+    $result = mysqli_query($conn, $sql);
+      if (mysqli_num_rows($result) == 0) { 
+              
+               }
+          else{
+            $row = mysqli_fetch_assoc($result);
+            $filename=$row["filename"];
+           
+            
+          }
+          return $filename ;
+}
+
+
 function delete_file($broadcastchannel ,$roomnumber)
 {
-  // check if single delete else incerease count
-  echo 'success';
+
+  $conn = mysqli_connect('localhost', 'root', '','instantsharing');
+  $sql = "DELETE FROM storage  WHERE broadcastchannel= '$broadcastchannel' and roomnumber = '$roomnumber'  and downloadcount >=1";
+     if (!mysqli_query($conn, $sql)) {
+        echo "deletion error".mysqli_error($conn);
+     }
 }
 
 
 function increasedownloadcount($filename,$broadcastchannel ,$roomnumber)
 {
-  // check if single delete else incerease count
-  delete_file($broadcastchannel ,$roomnumber);
-  echo 'success';
+
+  $conn = mysqli_connect('localhost', 'root', '','instantsharing');
+  $sql = "UPDATE storage  SET downloadcount = downloadcount +1 WHERE broadcastchannel= '$broadcastchannel' and roomnumber = '$roomnumber' and filename = '$filename'";
+   if (mysqli_query($conn, $sql)) {
+ 
+      delete_file($broadcastchannel ,$roomnumber);
+ 
+  }else{
+    echo "updation error".mysqli_error($conn);
+  }
 }
 
+function getshareablelink($filename,$broadcastchannel ,$roomnumber)
+{
+      $files = scandir("result");
+      for($a=1;$a<count($files);$a++)
+      {
+          if($files[$a] ===  $filename )
+          {           
+             $file = ("result/".$filename);
+             $url = 'thankyou/'.$file;
+            
+          }
+      }
 
+
+
+   
+}
 
 
 function savetodb($filename, $broadcastchannel ,$roomnumber,$sharingmode)
@@ -25,15 +73,18 @@ function savetodb($filename, $broadcastchannel ,$roomnumber,$sharingmode)
   if(isset($_SESSION['status'])){
     $broadcastchannel =  $_SESSION['broadcastchannel'];
   }
-  echo $broadcastchannel .$roomnumber.$sharingmode ;
+
+  $conn = mysqli_connect('localhost', 'root', '','instantsharing');
+  $sql = "INSERT INTO storage (filename, broadcastchannel,roomnumber,sharingmode)
+    VALUES ('$filename', '$broadcastchannel','$roomnumber', '$sharingmode')";
+
+    if (mysqli_query($conn, $sql)) {
+      
+    } else {
+        echo "Error: " . $sql . "<br>" . mysqli_error($conn);
+    }
 }
 
-function getFilename($broadcastchannel ,$roomnumber)
-{
-  echo $broadcastchannel;
-  $fname = '1583928424.zip';
-    return $fname ;
-}
 
 
 // BROADCASTING CONTENT
@@ -74,6 +125,7 @@ function getFilename($broadcastchannel ,$roomnumber)
                     $filename = $files['name'][0];
                     move_uploaded_file($tmp, 'result/'.$filename);
                     savetodb($filename ,$_POST['broadcastchannel'],$_POST['roomnumber'],$_POST['sharingmode']);
+                    getshareablelink($filename ,$_POST['broadcastchannel'],$_POST['roomnumber']);
           }
         }else{
           // error provide username
@@ -91,20 +143,26 @@ function getFilename($broadcastchannel ,$roomnumber)
 
 
 
-
 if(isset($_POST['download'])){
 
-      if(!empty($_POST['broadcastchannel']) && !empty($_POST['roomnumber']))
-    {
+      if(!empty($_POST['downloadbroadcastchannel']) && !empty($_POST['roomnumber']))
+    {    
+
       $filename = getFilename($_POST['downloadbroadcastchannel'],$_POST['roomnumber']);
+      
+      if(empty($filename))
+      {
+        header('Location:404');
+      }
       $files = scandir("result");
-      echo $filename;
+     
       for($a=1;$a<count($files);$a++)
       {
           if($files[$a] ===  $filename )
           {
             
-           increasedownloadcount($filename,$_POST['broadcastchannel'],$_POST['roomnumber']);     
+           increasedownloadcount($filename,$_POST['downloadbroadcastchannel'],$_POST['roomnumber']);     
+           
            $file = ("result/$filename");
            $filetype=filetype($file);
            $filename=basename($file);
